Implement new attach/detach Cinder flow
***************** WIP/POC ********************
Lots of details to address here, but this is a quick POC
to show a new workflow using Cinder create/remove attachment.

Goal of this patch was just to make this work for the standard
attach/detach path.  Still has a few places that need updated,
also we'll want to add some things on Cinder's side like a
refresh connection_info (rather than calling initialize_connection).

Currently we still use reserve etc, but we consolidate
initialize_connection and attach calls into a single
"create_attachment" call.

Same with terminate and destach, we combine those into a single
"remove_attachment" call.

Things that need to be done:
* Clearly there are quite a few places we still need to update (or we can leave
them calling the old mechanism).  The good thing about this approach is it's
not a BigBang change, we can keep using old methods at the same time.

* Take a look at the cleanup of connections (ie rm /dev/disk-by-path) and
  use either the return from list_attachments, or remove_attachment to
  determine if/when we should delete that connection (think multi-attach)

* API micro-versioning?

* Unit tests of course

Change-Id: Ifc01dbf98545104c998ab96f65ff8623a6db0f28
Depends-On: I90d54257862e1815a03ca8750040f7e4e6f35c0b
Depends-On: Ib4d229778b9f691ae86fca50059cca4c0958fcf8